backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.0
sqlalchemy==2.0.0
alembic==1.10.0
psycopg2-binary==2.9.6
pydantic==1.10.7
pydantic-settings==2.0.0
pydantic[email]==1.10.7
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.1.0
reportlab==4.0.4
apscheduler==3.10.4
python-dateutil==2.8.2
